requirements.txt
torch
torchvision
timm
opencv-python
numpy
pandas
scikit-learn
flask
tqdm
scipy
Pillow